Stop practice session when a user navigates from the screen
Fixes the problem, where the user navigates away from the practice screen still was participating in a practice.
Then the time runs out the user got redirected to the results screen.

Now the timer gets canceled when the practice screen loses its focus.
The practice session is restarted when the user navigates the screen again.